Module_1/requirements.txt
numpy
pandas
matplotlib
seaborn
Pillow
flake8
pytest
colorama

